requirements.txt
fastapi==0.128.7
uvicorn==0.40.0
python-dotenv==1.2.1
requests==2.32.5
httpx==0.28.1
feedparser==6.0.12
yfinance==1.1.0
google-genai==1.62.0
pydantic==2.12.5
jinja2==3.1.6
PyJWT==2.11.0
bcrypt==5.0.0
stripe==14.3.0
APScheduler==3.11.2
sentry-sdk[fastapi]==2.29.1
psycopg2-binary==2.9.10
redis==5.2.1
pyotp==2.9.0
qrcode[pil]==8.0
python-json-logger==3.2.1
authlib==1.4.1
itsdangerous==2.2.0
pytest==8.3.4
pytest-asyncio==0.25.0
psutil==7.0.0
